[REFACTOR] Phase 5: e2e reuses the app's flow decomposition (task 5.2)
Extract the pure nested->flat conversion (decomposeServerFlow and
helpers) from decodeFlow.ts into utils/serverFlowDecompose.ts - free of
React/rows imports so Playwright's Node process can load it - and
delete the near-verbatim 70-line copy in e2e/e2e.pw.ts. The e2e
fixtures' name fallbacks become a small explicit normalization step
(withFixtureNameFallbacks). Also drop e2e's hard-coded core service
UUID/resource literals in favor of evy-types/coreResources.

Verified: web build, lint, 155/155 unit, 5/5 e2e against the live
stack (dev DB reseeded after the run).
